+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/Gantt/BasicGanttCharts/Eventline.jsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import getGanttDefaultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gnatt/ganttDefault';
+import loadTracks from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/loadTracks';
+import tasksData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/tasksData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(getGanttDefaultConfig('#ganttEventline'));
+graphConfig.eventline = [
+  {
+    color: Carbon.helpers.COLORS.GREY,
+    style: {
+      strokeDashArray: '4,4',
+    },
+    value: new Date(2018, 10, 13).toISOString(),
+  },
+  {
+    color: Carbon.helpers.COLORS.BLACK,
+    style: {
+      strokeDashArray: '2,2',
+    },
+    value: new Date(2018, 10, 20).toISOString(),
+  },
+];
+graphConfig.showActionLegend = false;
+const dataset = loadTracks(tasksData, [], [], [], 4);
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <GanttChart graphID="ganttEventline" graphConfig={graphConfig} dataset={dataset} />
+  </React.Fragment>
+);