+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -7,3 +7,4 @@
   * Created this separate package for terra-dev-site and documentation.
   * Added data objects and examples for Line, Bar and Paired result graphs.
   * Added data objects and examples for Scatter graph.
+  * Added data objects and examples for Gantt chart.